updated examine_variance_behavior and deleted extra files.
